requirements.txt
asgiref==3.7.2
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0
django-import-export==3.3.4
django-wkhtmltopdf==3.4.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
fonttools==4.46.0
html5lib==1.1
idna==3.6
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.1.0
pycparser==2.21
pydyf==0.8.0
pyphen==0.14.0
pytz==2023.3.post1
PyYAML==6.0.1
reportlab==4.0.8
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tinycss2==1.2.1
typing_extensions==4.9.0
urllib3==2.1.0
weasyprint==60.2
webencodings==0.5.1
wkhtmltopdf==0.2
xlrd==2.0.1
XlsxWriter==3.1.9
xlwt==1.3.0
zopfli==0.2.3
